Exception handling

+++ src/main/java/com/devsuperior/dscommerce/services/ProductService.java
@@ -4,10 +4,15 @@ package com.devsuperior.dscommerce.services;
 import com.devsuperior.dscommerce.dto.ProductDTO;
 import com.devsuperior.dscommerce.entities.Product;
 import com.devsuperior.dscommerce.repositories.ProductRepository;
+import com.devsuperior.dscommerce.services.exceptions.DatabaseException;
+import com.devsuperior.dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
@@ -24,9 +29,13 @@ ProductService {
 
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id){
-        //Esse findById faz a consulta SQL no banco de dados
-        Product product = repository.findById(id).get();
-        return new ProductDTO(product);
+            //Esse findById faz a consulta SQL no banco de dados
+
+            Product product = repository.findById(id).orElseThrow(
+                    //Aqui você está lançando a própria exceção
+                    ()-> new ResourceNotFoundException("Recurso não encontrado"));
+            return new ProductDTO(product);
+
     }
 
     @Transactional(readOnly = true)
@@ -47,16 +56,28 @@ ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto){
-        //Vai estanciar pelo id do banco de dados
-        Product entity = repository.getReferenceById(id);
-        copyDTOToEntity(dto, entity);
-        return new ProductDTO(entity);
+        try{
+            //Vai estanciar pelo id do banco de dados
+            Product entity = repository.getReferenceById(id);
+            copyDTOToEntity(dto, entity);
+            return new ProductDTO(entity);
+        }catch (EntityNotFoundException e){
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id){
-        //Esse findById faz a consulta SQL no banco de dados
-        repository.deleteById(id);
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            //Esse findById faz a consulta SQL no banco de dados
+            repository.deleteById(id);
+        }catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
 
     private void copyDTOToEntity(ProductDTO dto, Product entity) {
